MCPJVM-410: normalize Feature Module internal types into models directories
