+++ tools/features/artifact-management/actions/types.ts
@@ -1,18 +1,5 @@
-import type { ArtifactManagementRequest } from "@tools-contracts/artifact-management";
-import type { ProbeRegistrySummary } from "@tools-core/probe-registry";
-
-export type ArtifactActionResult = {
-  content: Array<{ type: "text"; text: string }>;
-  structuredContent: Record<string, unknown>;
-};
-
-export type ArtifactActionContext = {
-  workspaceRootAbs: string;
-  getProbeRegistrySummary?: () => ProbeRegistrySummary | undefined;
-  reloadProbeRegistry?: () => ProbeRegistrySummary | undefined;
-};
-
-export type ArtifactActionRequest<T extends ArtifactManagementRequest["artifactType"]> = Extract<
-  ArtifactManagementRequest,
-  { artifactType: T }
->;
+export type {
+  ArtifactActionContext,
+  ArtifactActionRequest,
+  ArtifactActionResult,
+} from "../models/artifact_management.model";
