+++ tools/features/artifact-management/index.ts
@@ -2,4 +2,5 @@ export { artifactManagementDomain } from "./domain";
 export { buildFailClosedArtifactResponse } from "./shared/fail_closed";
 export { readProjectArtifact, writeProjectArtifact } from "./support/project_artifact_io";
 export type { ArtifactActionContext, ArtifactActionResult } from "./actions/types";
+export type { ArtifactActionRequest } from "./models/artifact_management.model";
 export type ArtifactManagementFeatureModule = "artifact-management";
