+++ tools/features/performance-suite/index.ts
@@ -2,3 +2,4 @@ export { executePerformanceRuntimeSuite } from "./actions/execute_performance_ru
 export { dispatchPerformanceSuiteAction } from "./actions/index";
 export { buildPerformanceMstaSummary } from "./performance_msta_summary";
 export type PerformanceSuiteFeatureModule = "performance-suite";
+export type * from "./models/performance_suite.model";
