+++ tools/features/route-synthesis/index.ts
@@ -1,7 +1,7 @@
 export { routeSynthesisDomain } from "./domain";
 export { discoverClassMethods, inferTargets } from "./shared/target_inference";
 export { generateRecipe } from "./shared/recipe_generation";
-export type { InferredTarget } from "./shared/target_inference";
+export type { ClassDiscoveryCandidate, ClassMethodSpan, InferredTarget } from "./models/target_inference.model";
 export { buildJavaIndex } from "./support/inference/java_index.util";
 export { enrichRuntimeCapture } from "./support/recipe_generate/runtime_capture_enrich.util";
 export {
